requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.7
django-allauth==0.40.0
django-cookies-samesite==0.8.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
filelock==3.0.12
idna==2.10
oauthlib==3.1.0
pipenv==2020.11.15
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
